requirements.txt
pandas
numpy
python-dateutil
streamlit
requests
python-dotenv
statsmodels
matplotlib
mplcyberpunk
pathlib
openpyxl
google-cloud-sql-connector
sqlalchemy